Fix some more compiler warnings
read_data_mutex was declared static in synchronizationlogging.h, thus
there was one copy of the variable per file. This was not intended.
Fixing that as well.

git-svn-id: svn://joust.ccs.neu.edu/home/tyler/svnroot/pthread-wrappers/fred@533 679428e0-fc1f-4e08-a6f4-36083810378a

+++ record-replay/pthreadwrappers.cpp
@@ -392,7 +392,7 @@ static int internal_pthread_create(pthread_t *thread,
 
     retval = _real_pthread_create(thread, &the_attr,
                                   start_wrapper, (void *)&createArg);
-    SET_COMMON2(my_entry, retval, (void*)retval);
+    SET_COMMON2(my_entry, retval, (void*)(unsigned long)retval);
     SET_COMMON2(my_entry, my_errno, errno);
 
     waitForChildThreadToInitialize();
